+++ src/main/java/Shared/AttackOrder.java
@@ -18,12 +18,12 @@ public class AttackOrder extends OrderBasic {
         int att_units = units, def_units = toT.getUnits();
         Player attacker = this.getPlayer(), defender = toT.getOwner();
         while(att_units > 0 && def_units > 0) {
-//            int attNum = rollDice(1, 20);
-//            int defNum = rollDice(1, 20);
-//            if (attNum > defNum) def_units--;
-//            else att_units--;
+            int attNum = rollDice(1, 20);
+            int defNum = rollDice(1, 20);
+            if (attNum > defNum) def_units--;
+            else att_units--;
 
-            // for accurate testing, we don't use random number now
+            /* For accurate testing, we don't use random number
             if (att_units > def_units) {
                 att_units -= def_units;
                 def_units = 0;
@@ -36,11 +36,11 @@ public class AttackOrder extends OrderBasic {
                 def_units = 1;
                 att_units = 0;
             }
-            // ------------------------------------------------------
+            */
+
         }
         if(att_units > 0) {  // attacker wins
             toT.setUnits(att_units);
-            System.out.println("set " + toT.getName() + "'s owner to " + attacker.getName());
             toT.setOwner(attacker);
         }
         else {  //defender wins
