+++ src/main/java/Shared/OrderBasic.java
@@ -2,7 +2,6 @@ package Shared;
 
 public abstract class OrderBasic extends Order {
     protected Player player;
-    protected String orderType;
     protected Territory fromT;
     protected Territory toT;
     protected int units;
